+++ tests/test_smartbridge.py
@@ -2,7 +2,7 @@
 import asyncio
 from collections import defaultdict
 from datetime import timedelta
-import json
+import orjson
 import logging
 import os
 import re
@@ -54,7 +54,7 @@ def response_from_json_file(filename: str) -> Response:
     """Fetch a response from a saved JSON file."""
     responsedir = os.path.join(os.path.split(__file__)[0], "responses")
     with open(os.path.join(responsedir, filename), "r", encoding="utf-8") as ifh:
-        return Response.from_json(json.load(ifh))
+        return Response.from_json(orjson.loads(ifh.read()))
 
 
 class Request(NamedTuple):
@@ -280,7 +280,7 @@ class Bridge:
 
         # Check the zone status on each zone
         requested_zones = []
-        for _ in range(0, 4):
+        for _ in range(0, 5):
             request, response = await wait(leap.requests.get())
             logging.info("Read %s", request)
             assert request.communique_type == "ReadRequest"
@@ -308,6 +308,7 @@ class Bridge:
             "/zone/1/status",
             "/zone/2/status",
             "/zone/3/status",
+            "/zone/4/status",
             "/zone/6/status",
         ]
 
@@ -838,6 +839,21 @@ async def test_device_list(bridge: Bridge):
             "tilt": None,
             "occupancy_sensors": None,
         },
+        "11": {
+            "area": "3",
+            "device_id": "11",
+            "device_name": "WoodBlinds",
+            "name": "Living Room_WoodBlinds",
+            "type": "Tilt",
+            "zone": "4",
+            "model": "TLT-EDU-B-J",
+            "serial": 4569,
+            "current_state": -1,
+            "fan_speed": None,
+            "button_groups": None,
+            "tilt": None,
+            "occupancy_sensors": None,
+        },
     }
 
     assert devices == expected_devices
@@ -875,6 +891,19 @@ async def test_device_list(bridge: Bridge):
             Body={"ZoneStatus": {"Tilt": 25, "Zone": {"href": "/zone/3"}}},
         )
     )
+
+    bridge.leap.send_unsolicited(
+        Response(
+            CommuniqueType="ReadResponse",
+            Header=ResponseHeader(
+                MessageBodyType="OneZoneStatus",
+                StatusCode=ResponseStatus(200, "OK"),
+                Url="/zone/4/status",
+            ),
+            Body={"ZoneStatus": {"Tilt": 40, "Zone": {"href": "/zone/4"}}},
+        )
+    )
+
     devices = bridge.target.get_devices()
     assert devices["2"]["current_state"] == 100
     assert devices["2"]["fan_speed"] is None
@@ -882,6 +911,8 @@ async def test_device_list(bridge: Bridge):
     assert devices["3"]["fan_speed"] == FAN_MEDIUM
     assert devices["10"]["current_state"] == -1
     assert devices["10"]["tilt"] == 25
+    assert devices["11"]["current_state"] == -1
+    assert devices["11"]["tilt"] == 40
 
     devices = bridge.target.get_devices_by_domain("light")
     assert len(devices) == 1
@@ -906,10 +937,13 @@ async def test_device_list(bridge: Bridge):
     assert devices[0]["device_id"] == "3"
 
     devices = bridge.target.get_devices_by_domain("cover")
-    assert [device["device_id"] for device in devices] == ["7", "10"]
+    assert [device["device_id"] for device in devices] == ["7", "10", "11"]
 
     devices = bridge.target.get_devices_by_type("SerenaTiltOnlyWoodBlind")
     assert [device["device_id"] for device in devices] == ["10"]
+
+    devices = bridge.target.get_devices_by_type("Tilt")
+    assert [device["device_id"] for device in devices] == ["11"]
 
 
 @pytest.mark.asyncio
@@ -2362,6 +2396,80 @@ async def test_qsx_set_ketra_level_with_fade(qsx_processor: Bridge):
             "Command": {
                 "CommandType": "GoToSpectrumTuningLevel",
                 "SpectrumTuningLevelParameters": {"Level": 50, "FadeTime": "00:00:04"},
+            }
+        },
+    )
+    qsx_processor.leap.requests.task_done()
+    task.cancel()
+    await qsx_processor.target.close()
+
+
+@pytest.mark.asyncio
+async def test_qsx_set_LumarisRGB_color(qsx_processor: Bridge):
+    """
+    Test that setting the color of a Lumaris RGB produces the
+    right command.
+    """
+    hue = 150
+    saturation = 30
+    full_color = color_value.FullColorValue(hue, saturation)
+    task = asyncio.get_running_loop().create_task(
+        qsx_processor.target.set_value("991", color_value=full_color)
+    )
+    command, _ = await qsx_processor.leap.requests.get()
+    assert command == Request(
+        communique_type="CreateRequest",
+        url="/zone/991/commandprocessor",
+        body={
+            "Command": {
+                "CommandType": "GoToSpectrumTuningLevel",
+                "SpectrumTuningLevelParameters": {
+                    "ColorTuningStatus": {
+                        "HSVTuningLevel": {"Hue": hue, "Saturation": saturation}
+                    }
+                },
+            }
+        },
+    )
+    qsx_processor.leap.requests.task_done()
+    task.cancel()
+
+    kelvin = 2700
+    warm_color = color_value.WarmCoolColorValue(kelvin)
+    task = asyncio.get_running_loop().create_task(
+        qsx_processor.target.set_value("991", color_value=warm_color)
+    )
+    command, _ = await qsx_processor.leap.requests.get()
+    assert command == Request(
+        communique_type="CreateRequest",
+        url="/zone/991/commandprocessor",
+        body={
+            "Command": {
+                "CommandType": "GoToSpectrumTuningLevel",
+                "SpectrumTuningLevelParameters": {
+                    "ColorTuningStatus": {"WhiteTuningLevel": {"Kelvin": kelvin}}
+                },
+            }
+        },
+    )
+    qsx_processor.leap.requests.task_done()
+    task.cancel()
+
+    task = asyncio.get_running_loop().create_task(
+        qsx_processor.target.set_warm_dim("991", True)
+    )
+    command, _ = await qsx_processor.leap.requests.get()
+    assert command == Request(
+        communique_type="CreateRequest",
+        url="/zone/991/commandprocessor",
+        body={
+            "Command": {
+                "CommandType": "GoToSpectrumTuningLevel",
+                "SpectrumTuningLevelParameters": {
+                    "ColorTuningStatus": {
+                        "CurveDimming": {"Curve": {"href": "/curve/1"}}
+                    }
+                },
             }
         },
     )