+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.23.0] - 2025-01-05
+
+### Added
+
+- Support for Lumaris RGB + Tunable White Tape Light.
+- Support for wood tilt blinds (Sivoia?).
+
 ## [0.22.0] - 2024-10-04
 
 ### Added
@@ -330,7 +337,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Ability to interact with Caséta bridge using LIP over Telnet and LEAP over SSH.
 
-[unreleased]: https://github.com/gurumitts/pylutron-caseta/compare/v0.22.0...HEAD
+[unreleased]: https://github.com/gurumitts/pylutron-caseta/compare/v0.23.0...HEAD
+[0.23.0]: https://github.com/gurumitts/pylutron-caseta/compare/v0.22.0...v0.23.0
 [0.22.0]: https://github.com/gurumitts/pylutron-caseta/compare/v0.21.1...v0.22.0
 [0.21.1]: https://github.com/gurumitts/pylutron-caseta/compare/v0.21.0...v0.21.1
 [0.21.0]: https://github.com/gurumitts/pylutron-caseta/compare/v0.20.0...v0.21.0